+++ apps/www/src/app/[lang]/blog/[slug]/page.tsx
@@ -1,0 +1,2 @@
+export { default } from '@/modules/blog/page/detail';
+export * from '@/modules/blog/page/detail';