+++ apps/www/src/modules/blog/utilities/date.ts
@@ -1,0 +1,4 @@
+export const formatDate = (dateString: string) => {
+    const options: Intl.DateTimeFormatOptions = { year: "numeric", month: "long", day: "numeric" }
+    return new Date(dateString).toLocaleDateString(undefined, options)
+}