+++ apps/www/src/modules/common/types/common.d.ts
@@ -7,6 +7,7 @@ export interface Footer {
 }
 
 export interface ProjectsAttributes extends BaseServerAttributes {
+  slug: any;
   title: string;
   summary: string;
   locale: string;
@@ -38,12 +39,17 @@ export interface Services {
 }
 
 export interface BlogAttributes extends BaseServerAttributes {
+  author: StrapiResponse<TeamMember>;
   title: string;
   summary: string;
   body?: Array<ComponentsOptions>;
+  categories: StrapiResponse<Array<CategoryAttributes>>;
   image: MediaResponse;
+  slug: string;
   seo: Seo;
   locale: string;
+  createdAt: string;
+  updatedAt: string;
 }
 
 export interface Blogs {
@@ -176,4 +182,15 @@ export interface Seo {
   description: string;
   keywords: string;
   preventIndexing: boolean;
+}
+
+export interface Category {
+  name: string;
+  createdAt: string;
+  updatedAt: string;
+}
+
+export interface CategoryAttributes {
+  id: number;
+  attributes: Category;
 }