+++ apps/www/src/modules/work/actions/work.action.ts
@@ -1,50 +1,42 @@
 /* eslint-disable turbo/no-undeclared-env-vars */
 "use server";
 
-import { config } from "dotenv";
-
-import qs from "qs";
-import { AppConfig } from "@/config/app.config";
+import { baseFetch } from "@/modules/common/utilities/fetch";
 import {
   ProjectsServerResponse,
   ProjectServerResponse,
 } from "../types/response";
 
-config();
+const fetchData = function <T>(filter?: { [key: string]: string | number }) {
+  const populate = [
+    "image",
+    "description",
+    "description.badges",
+    "description.image",
+    "description.image.image",
+    "description.heading",
+    "description.overview",
+    "seo",
+  ];
 
-async function baseFetch<TData>(id?: string | number) {
-  const query = qs.stringify({
-    populate: [
-      "description",
-      "description.badges",
-      "description.image",
-      "description.image.image",
-      "description.heading",
-      "seo",
-    ],
+  return baseFetch<T>({
+    entity: "projects",
+    populate,
+    by: filter,
   });
-  const baseUrl = `${AppConfig.strapi.url}/api/projects`;
-  const url = id ? `${baseUrl}/${id}?${query}` : `${baseUrl}?${query}`;
-
-  const request = await fetch(url, {
-    method: "GET",
-    headers: {
-      Authorization: `Bearer ${process.env.STRAPI_ACCESS_TOKEN}`,
-    },
-    next: { revalidate: 120 },
-  });
-
-  const data = await request.json();
-
-  return data as TData;
-}
+};
 
 export async function getAllWorkAction() {
-  const data = baseFetch<ProjectsServerResponse>();
+  const data = fetchData<ProjectsServerResponse>();
   return data;
 }
 
 export async function getWorkAction(id: string | number) {
-  const data = baseFetch<ProjectServerResponse>(id);
+  const data = fetchData<ProjectServerResponse>({ id });
+  return data;
+}
+
+export async function getWorkBySlugAction(slug: string) {
+  const data = fetchData<ProjectServerResponse>({ slug });
   return data;
 }